swiftui: simplify the UI
