+++ src/frontend/swiftui/ContentView.swift
@@ -1,15 +1,10 @@
 import SwiftUI
 
 struct ContentView: View {
-    @State private var selection: SidebarSelectionState? = .gamelist
-    
     var body: some View {
-        NavigationSplitView(sidebar: {
-            SidebarView(selection: $selection)
-                .frame(minWidth: 180)
-        }, detail: {
-            DetailView(selection: $selection)
-        })
+        NavigationStack {
+            GameListView()
+        }
         .toolbar {
             ToolbarItems()
         }
