+++ src/frontend/swiftui/EmulationView.swift
