--- src/frontend/swiftui/Sidebar/NowPlayingView.swift
@@ -1,9 +1,0 @@
-import SwiftUI
-
-struct NowPlayingView: View {
-    var body: some View {
-        Image(systemName: "play.circle.fill")
-            .resizable()
-            .frame(width: 100, height: 100)
-    }
-}
