+++ src/frontend/swiftui/ToolbarItems.swift
@@ -10,7 +10,10 @@ struct ToolbarItems: ToolbarContent {
                 isFilePickerPresented.toggle()
             }
             // TODO: change allowedContentTypes to nsp etc
-            .fileImporter(isPresented: $isFilePickerPresented, allowedContentTypes: [.data], allowsMultipleSelection: false) { result in
+            .fileImporter(
+                isPresented: $isFilePickerPresented, allowedContentTypes: [.data],
+                allowsMultipleSelection: false
+            ) { result in
                 switch result {
                 case .success(let fileurls):
                     print(fileurls.count)
@@ -23,15 +26,6 @@ struct ToolbarItems: ToolbarContent {
                     print(error)
                 }
             }
-
-            Button("Reset", systemImage: "power") { }
-
-            Button("Save States", systemImage: "photo.on.rectangle.angled") { }
-
-            Button("Controller 1", systemImage: "gamecontroller") { }
-
-            Button("Controller 2", systemImage: "gamecontroller") { }
-
         }
     }
